user [nfc]: Inline "card" into trivial "screen" wrapper UsersScreen
